Rename Test directory, add version no, cleanup file headers.
Renamed the "Test" directory to "tests". The documentation of options in
the PHPDoc of WikiCreole.php was removed and replaced with a reference
to the README file. All of the file headers were cleaned up, moving the
PHPDoc tags to the class level docblock instead of the file level docblock.
Finally the WikiCreole class (and thus the library) got a version number
in anticipation of using with Composer/Packagist. The output of the
"FullTest" was added to the .gitignore file. It should not be tracked b/c
it is subject to change anytime tests are run.